1.优化配置项 2.添加Code类型

+++ markdown/src/main/java/com/star/plugin/markdown/span/HSpan.java
@@ -15,9 +15,9 @@ import androidx.annotation.NonNull;
 public class HSpan extends MetricAffectingSpan implements MarkDownSpan {
 
     private int color;
-    private int size;
+    private float size;
 
-    public HSpan(int color, int size) {
+    public HSpan(int color, float size) {
         this.color = color;
         this.size = size;
     }
